Refactoring: see Description
- changed Button Style in Encryption and Decryption Page
- added summaries to public methods
- moved methods from CryptoClass (PresentationLayer) into
AesGcmCryptoLibrary (BusinessLayer)
