+++ BusinessLayer/EP_HSRlearnIT.BusinessLayer.csproj
@@ -35,6 +35,9 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="Castle.Core">
+      <HintPath>..\packages\Castle.Core.3.3.3\lib\net45\Castle.Core.dll</HintPath>
+    </Reference>
     <Reference Include="Security.Cryptography">
       <HintPath>..\externalLibraries\Security.Cryptography.dll</HintPath>
     </Reference>
