+++ PresentationLayer/Exercises/CryptoClass.cs
@@ -2,10 +2,8 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using System.Windows.Controls;
 using System.Windows.Media;
-using Castle.Core.Internal;
 using EP_HSRlearnIT.BusinessLayer.CryptoTools;
 using EP_HSRlearnIT.BusinessLayer.UniversalTools;
 
@@ -27,6 +25,11 @@ namespace EP_HSRlearnIT.PresentationLayer.Exercises
         #endregion
 
         #region Public Methods
+        /// <summary>
+        /// This event is used to convert the TextBox input into Hex Values and update the correspondig HexTextBox.
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         public void TextBox_TextChanged(object sender, TextChangedEventArgs e)
         {
             //Get control that raised this event
@@ -42,6 +45,12 @@ namespace EP_HSRlearnIT.PresentationLayer.Exercises
             }
         }
 
+        /// <summary>
+        /// This Method is used to convert the TextBox input into Hex Values and update the correspondig HexTextBox.
+        /// It can be called, if there is no way to register an event.
+        /// </summary>
+        /// <param name="values"></param>
+        /// <param name="hexBox"></param>
         public void ChangeHexBox(char[] values, TextBox hexBox)
         {
             string hexOutput = "";
@@ -62,6 +71,12 @@ namespace EP_HSRlearnIT.PresentationLayer.Exercises
             }
         }
 
+        /// <summary>
+        /// This event is used to convert the HexTextBox input into chars and update the corresponding TextBox.
+        /// If a non hex value is entered a warning will be shown.
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         public void HexTextBox_TextChanged(object sender, TextChangedEventArgs e)
         {
             string textOutput = "";
@@ -73,19 +88,19 @@ namespace EP_HSRlearnIT.PresentationLayer.Exercises
                 // get the name of the corresponding hexBlock. The Box suffix (3 chars) is removed and Block is appended. 
                 string nameHexBlock = nameHexTextBox.Substring(0, nameHexTextBox.Length - 3) + "Block";
 
-                //Get control that will be updated
-                string nameTextField = nameHexTextBox.Substring(3, nameHexTextBox.Length - 3);
-                TextBox textBox = (TextBox)FindName(nameTextField);
+                //Get controls that will be updated
+                string nameTextBox = nameHexTextBox.Substring(3, nameHexTextBox.Length - 3);
+                TextBox textBox = (TextBox)FindName(nameTextBox);
                 TextBlock textBlock = (TextBlock)FindName(nameHexBlock);
                 if (textBlock == null) { return; }
 
-                string textBlockAnzeige = textBlock.Text;
-
-                if (textBlockAnzeige.Contains(" Ungültige Eingabe!"))
+                //check if the text was set to " Ungültige Eingabe!" in an earlier call of this method and reverse it 
+                string textBlockName = textBlock.Text;
+                if (textBlockName.Contains(" Ungültige Eingabe!"))
                 {
                     //19 is the length of suffix text " Ungültige Eingabe!"
-                    textBlockAnzeige = textBlockAnzeige.Substring(0, textBlockAnzeige.Length - 19);
-                    textBlock.Text = textBlockAnzeige;
+                    textBlockName = textBlockName.Substring(0, textBlockName.Length - 19);
+                    textBlock.Text = textBlockName;
                     textBlock.Foreground = Brushes.Black;
                 }
 
@@ -96,14 +111,16 @@ namespace EP_HSRlearnIT.PresentationLayer.Exercises
 
                 string hexValue = hexTextBox.Text;
 
+                //if the input is invalid the following will be executed
                 if (!IsHex(hexValue))
                 {
-                    textBlock.Text = textBlockAnzeige + " Ungültige Eingabe!";
+                    textBlock.Text = textBlockName + " Ungültige Eingabe!";
                     textBlock.Foreground = Brushes.Red;
                     textBox.TextChanged += TextBox_TextChanged;
                     return;
                 }
 
+                
                 ArrayList list = new ArrayList();
                 while (hexValue.Length > 1)
                 {
@@ -124,37 +141,11 @@ namespace EP_HSRlearnIT.PresentationLayer.Exercises
             }
         }
 
-        public void GenerateHexKey(string key, TextBox hexPasswordBox)
-        {
-            byte[] keyArray = StringToBytes(key);
-            int keySize = keyArray.Length;
-
-            IEnumerable<byte> bigKey = keyArray;
-
-            if (keySize < 32)
-            {
-                for (int i = 1; i <= 32 / keySize; i++)
-                {
-                    bigKey = bigKey.Concat(keyArray);
-                }
-            }
-
-            bigKey = bigKey.Take(32);
-            byte[] result = new byte[32];
-            int counter = 0;
-
-            bigKey.ForEach(i =>
-            {
-                byte b = i;
-                result[counter] = b;
-                counter++;
-            });
-
-            string bigKeyString = BytesToString(result);
-
-            ChangeHexBox(bigKeyString.ToCharArray(), hexPasswordBox);
-        }
-
+        /// <summary>
+        /// This event is used to save the Progress of a HexTextBox.
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         public void HexTextBox_LostKeyboardFocus(object sender, System.Windows.Input.KeyboardFocusChangedEventArgs e)
         {
             TextBox source = e.Source as TextBox;
@@ -165,6 +156,11 @@ namespace EP_HSRlearnIT.PresentationLayer.Exercises
             }
         }
 
+        /// <summary>
+        /// This event is used to save the Progress of a TextBox.
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         public void TextBox_LostKeyboardFocus(object sender, System.Windows.Input.KeyboardFocusChangedEventArgs e)
         {
             TextBox source = e.Source as TextBox;
@@ -178,62 +174,12 @@ namespace EP_HSRlearnIT.PresentationLayer.Exercises
 
         }
 
-        public static byte[] HexStringToByteArray(string hex)
-        {
-            return Enumerable.Range(0, hex.Length)
-                             .Where(x => x % 2 == 0)
-                             .Select(x => Convert.ToByte(hex.Substring(x, 2), 16))
-                             .ToArray();
-        }
-
-        public string BytesToString(byte[] bytes)
-        {
-            char[] chars = new char[bytes.Length];
-
-            StringBuilder sb = new StringBuilder();
-            int i = 0;
-            foreach (byte b in bytes)
-            {
-                chars[i] = Convert.ToChar(b);
-                sb.Append(chars[i]);
-                i++;
-            }
-            return sb.ToString();
-        }
-
-        public byte[] StringToBytes(string toConvert)
-        {
-            byte[] bytes = new byte[toConvert.Length];
-
-            int i = 0;
-            foreach (char c in toConvert)
-            {
-                bytes[i] = Convert.ToByte(c);
-                i++;
-            }
-
-            return bytes;
-        }
         #endregion
 
         #region Private Methods
-        //TODO mit Regex vereinfachbar?
         private bool IsHex(IEnumerable<char> chars)
         {
-            //string check = Convert.ToString(chars);
-            //return System.Text.RegularExpressions.Regex.IsMatch(check, @"\A\b[0-9a-fA-F]+\b\Z");
-            foreach (var c in chars)
-            {
-                bool isHex = ((c >= '0' && c <= '9') ||
-                              (c >= 'a' && c <= 'f') ||
-                              (c >= 'A' && c <= 'F'));
-
-                if (!isHex)
-                {
-                    return false;
-                }
-            }
-            return true;
+            return chars.Select(c => (c >= '0' && c <= '9') || (c >= 'a' && c <= 'f') || (c >= 'A' && c <= 'F')).All(isHex => isHex);
         }
 
         private void SaveProgressHelper(TextBox source, string elementName)
