+++ PresentationLayer/Exercises/DecryptionPage.xaml
@@ -28,14 +28,14 @@
             <RowDefinition Height="Auto" Name="Row7"/>
             <RowDefinition Height="Auto" Name="Row8"/>
             <RowDefinition Height="Auto" Name="Row9"/>
-            <RowDefinition Height="Auto" Name="Row10"/>
+            <RowDefinition Height="0.75*" Name="Row10"/>
             <RowDefinition Height="Auto" Name="Row11"/>
             <RowDefinition Height="Auto" Name="Row12"/>
             <RowDefinition Height="*" Name="Row13"/>
         </Grid.RowDefinitions>
         <Border CornerRadius="6" BorderBrush="Gray"  Background="LightGray" BorderThickness="2" Grid.ColumnSpan="1" Grid.RowSpan="14" Margin="0,0,0,10"/>
 
-        <Button Grid.Row="0" Grid.Column="0" Click="OnImportButtonClick" Content="File Importieren" Name="ImportButton" Height="40" Width="90" HorizontalAlignment="Left" Margin="15,10,15,2"/>
+        <Button Grid.Row="0" Grid.Column="0" Click="OnImportButtonClick" Content="File Importieren" Name="ImportButton" Style="{StaticResource ResourceKey=ProgramButtonStyle}" Height="40" Width="95" HorizontalAlignment="Left" Margin="15,10,15,2"/>
         <TextBlock Grid.Row="1" Grid.Column="0" Text="Verschlüsselter Text:" Margin="15,2,15,0" Name="CipherTextBlock"/>
         <TextBox Grid.Row="2" Grid.Column="0" Name="CipherTextBox" Margin="15,2,15,5" TextChanged="TextBox_TextChanged" LostKeyboardFocus="TextBox_LostKeyboardFocus" TextWrapping="Wrap" ScrollViewer.VerticalScrollBarVisibility="Auto"/>
         <TextBlock Grid.Row="3" Grid.Column="0" Text="Additional Authenticated Data: (optional)" Name="AadBlock" Margin="15,2,15,0" HorizontalAlignment="Left"/>
@@ -75,7 +75,7 @@
         </Image>
         <TextBox Grid.Row="10" Grid.Column="0" Name="DecryptionPasswortBox" Margin="15,2,15,5" TextChanged="TextBox_TextChanged" TextWrapping="Wrap"/>
         <DockPanel Grid.Row="11" Grid.Column="0" Margin="15,2,15,2">
-            <Button Click="OnDecryptionButtonClick" Content="Entschlüsseln" Name="DecryptionButton" Height="40" Width="85" Margin="0,10,7.5,10">
+            <Button Click="OnDecryptionButtonClick" Content="Entschlüsseln" Name="DecryptionButton" Style="{StaticResource ResourceKey=ProgramButtonStyle}" Height="40" Width="85" Margin="0,10,7.5,10">
                 <Button.IsEnabled>
                     <MultiBinding Converter="{StaticResource FormFilledDecryption}">
                         <Binding Path="Text.Length" ElementName="HexCipherTextBox"/>
@@ -86,7 +86,7 @@
                     </MultiBinding>
                 </Button.IsEnabled>
             </Button>
-            <Button Click="OnEncryptionButtonClick" Content="Verschlüsseln" Name="EncrpytionButton" Height="40" Width="85" HorizontalAlignment="Left" Margin="7.5,10,0,10" IsEnabled="{Binding ElementName=PlainTextBox, Path=Text.Length}"/>
+            <Button Click="OnEncryptionButtonClick" Content="Verschlüsseln" Name="EncrpytionButton" Style="{StaticResource ResourceKey=ProgramButtonStyle}" Height="40" Width="85" HorizontalAlignment="Left" Margin="7.5,10,0,10" IsEnabled="{Binding ElementName=PlainTextBox, Path=Text.Length}"/>
         </DockPanel>
         <TextBlock Grid.Row="12" Grid.Column="0" Text="Klartext:" Name="PlainTextBlock" Margin="15,2,15,0"/>
         <TextBox Grid.Row="13" Grid.Column="0" Name="PlainTextBox" Margin="15,2,15,18" TextChanged="TextBox_TextChanged" IsReadOnly="True" TextWrapping="Wrap" ScrollViewer.VerticalScrollBarVisibility="Auto"/>
